Change thesis to be approach oriented
Previously, my thesis described my goals for the work focused around
Holmes and its target capabilities.
I've changed it to describe the general idea of using a logic language
to get better results and easier integration, and left Holmes as my way
of proving it.
Hopefully this approach will allow for greater flexibility in the actual
work when it commences and an easier defense when it finishes.